Crate rust_test

Modules

atoi

Implement the myAtoi(string s) function, which converts a string to a 32-bit signed integer (similar to C/C++’s atoi function).

palindrome_sub

Given a string s, return the longest palindromic substring in s.

power_of_four

Given an integer n, return true if it is a power of four. Otherwise, return false. An integer n is a power of four, if there exists an integer x such that n == 4^x.

power_of_two

Given an integer n, return true if it is a power of two. Otherwise, return false. An integer n is a power of two, if there exists an integer x such that n == 2x.

remove_nth_from_end

Given the head of a linked list, remove the nth node from the end of the list and return its head.

roman_to_integer

Given a roman numeral, convert it to an integer.

rotate_image

You are given an n x n 2D matrix representing an image, rotate the image by 90 degrees (clockwise). You have to rotate the image in-place, which means you have to modify the input 2D matrix directly. DO NOT allocate another 2D matrix and do the rotation.